refactor: extract 3 helper structs from TranscriptionClient (#648 slices B/C/D)
Bundles 3 helper-struct extractions from TranscriptionClient.swift since they
all touch the same source file and 2 of them share a private response type:

**Slice B (#648 B)** — MultipartFormDataBuilder.swift:
- MultipartFormDataBuilder (34 lines)
- SHA-256: 0e617c16b3f3d187679f689debec88b8657f1546e52c5b9c70641bf35a0513a1

**Slice C (#648 C)** — TranscriptionRequestFactory.swift:
- TranscriptionRequestFactory (38 lines)
- SHA-256: a59951643923a9a8db5dc722c43734b188a46c61943e44ae4ecec83f8f79f4aa

**Slice D (#648 D)** — VerboseTranscriptionResponseParser.swift:
- VerboseTranscriptionResponseParser (18 lines)
- Plus the private VerboseTranscriptionResponse type it decodes into (was
  file-private in TranscriptionClient.swift; moves with the parser to keep the
  private scope intact — no visibility widen)
- SHA-256 (parser): b9dbd1425f0fbe0f11f40ac4f7f08208cc5a9f5172b234e1e20e1bd6362c6e75

Justification for bundling per PR #646 review 39: normally sequential PRs are
the discipline, but here slice D's parser depends on the private
VerboseTranscriptionResponse struct — moving them together in one PR keeps
the private-scope invariant without requiring a `internal` visibility widen
in an intermediate state. Slices B and C are bundled for convenience since
they touch the same file and land in the same sequence.

Reduces TranscriptionClient.swift from 711 to 618 lines (post-#654 baseline was
711 lines after the 3 model structs moved).

Tests: 667 (unchanged).
